+++ src/main/java/game/state/battle/entity/components/VitalsComponent.java
@@ -35,8 +35,8 @@ public class VitalsComponent implements RenderableComponent {
         renderer.setColor(healthColor);
         int healthWidth = (int) ((32 - 10) * healthPercentage);
         int healthHeight = 5;
-        int healthX = (position.getX() * 32) + 5;
-        int healthY = (position.getY() * 32) + 32 - 5;
+        int healthX = (int) ((position.getX() * 32) + 5);
+        int healthY = (int) ((position.getY() * 32) + 32 - 5);
 
         renderer.setColor(Color.BLACK);
         renderer.fillRect(healthX, healthY, 32 - 10, healthHeight);
